+++ src/23table.js
@@ -9,21 +9,29 @@
 // Table class
 var Table = alasql.Table = function(params){
 
-	// Columns
-	this.columns = [];
-	this.xcolumns = {};
-	// Data array
+	// Step 1: Data array
 	this.data = [];
 
+	// Step 2: Columns
+	this.columns = [];
+	this.xcolumns = {};
+
+	// Step 3: indices
 	this.inddefs = {};
 	this.indices = {};
-
 	this.uniqs = {};
 	this.uniqdefs = {};	
 
+	// Step 4: identities
 	this.identities = {};
-	this.checkfn = [];
 
+	// Step 5: checkfn...
+	this.checkfn = [];
+	this.checkfns = []; // For restore... to be done...
+
+	// Step 6: INSERT/DELETE/UPDATE
+
+	// Step 7: Triggers...
 	// Create trigger hubs
 	this.beforeinsert = {};
 	this.afterinsert = {};
@@ -37,6 +45,7 @@ var Table = alasql.Table = function(params){
 	this.afterupdate = {};
 	this.insteadofupdate = {};
 
+	// Done
 	extend(this,params);
 };