+++ src/40select.js
@@ -155,6 +155,7 @@ yy.Select.prototype.compile = function(databaseid) {
 
 	// Array with columns to be removed
     query.removeKeys = [];
+    query.aggrKeys = [];
 
 	query.explain = this.explain; // Explain
 	query.explaination = [];
@@ -215,6 +216,11 @@ yy.Select.prototype.compile = function(databaseid) {
 	}
 
 
+	// 8. Compile ORDER BY clause
+	if(this.order){
+		query.orderfn = this.compileOrder(query);
+	}
+
 	if(this.group || query.selectGroup.length>0) {
 		query.selectgfn = this.compileSelectGroup2(query);
 	} else {
@@ -225,10 +231,6 @@ yy.Select.prototype.compile = function(databaseid) {
 	// 7. Compile DISTINCT, LIMIT and OFFSET
 	query.distinct = this.distinct;
 
-	// 8. Compile ORDER BY clause
-	if(this.order){
-		query.orderfn = this.compileOrder(query);
-	}
 
 	// 9. Compile PIVOT clause
 	if(this.pivot) query.pivotfn = this.compilePivot(query);