+++ test/test168.js
@@ -11,7 +11,7 @@ if(typeof exports === 'object') {
 describe('Test 168a - read XLSX', function() {
 	this.timeout(10000);
 	it("1. Read XLSX file", function(done) {
-		var res = alasql('select * from xlsx("'+dirname+'/test168.xlsx")',[],function(res){
+		var res = alasql('select * from xlsx("'+dirname+'/test168.xlsx",{headers:false})',[],function(res){
 			assert(res.length == 6);
 //			console.log(res);
 			done();		
@@ -49,7 +49,7 @@ describe('Test 168a - read XLSX', function() {
 describe('Test 168b - read XLS', function() {
 	this.timeout(9000);
 	it("1. Read XLS file", function(done) {
-		var res = alasql('select * from xls("'+dirname+'/test168.xls")',[],function(res){
+		var res = alasql('select * from xls("'+dirname+'/test168.xls",{headers:false})',[],function(res){
 			assert(res.length == 6);
 //			console.log(res);
 			done();